Unify data type of some IDs to uint_32t.
Currently some tests will fail because they use large numbers
intentionally.

+++ server/src/UnifiedDataStore.cc
@@ -141,7 +141,7 @@ struct UnifiedDataStore::PrivateContext
 			if (targetServerId != ALL_SERVERS) {
 				const MonitoringServerInfo &info
 					= arm->getServerInfo();
-				if (static_cast<int>(targetServerId) != info.id) {
+				if (targetServerId != info.id) {
 					remainingArmsCount--;
 					continue;
 				}
